refactor: 메서드를 목적에 맞게 이동
- HttpRequest의 readFile 메서드를 HttpResponse로 이동 시키고,
  RequestHandler에 변경 사항을 반영했습니다.
